exception handling and response code of crud operations added+fixed

+++ src/main/java/fi/devolon/demo/service/CompanyService.java
@@ -1,6 +1,7 @@
 package fi.devolon.demo.service;
 
 import fi.devolon.demo.exceptions.CompanyNotFoundException;
+import fi.devolon.demo.exceptions.MissingValueException;
 import fi.devolon.demo.model.Company;
 import fi.devolon.demo.repository.CompanyRepository;
 import lombok.AllArgsConstructor;
@@ -8,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
-
 @Service
 @AllArgsConstructor(onConstructor = @__(@Autowired))
 public class CompanyService implements BasicRestService<Company>{
@@ -18,6 +18,11 @@ public class CompanyService implements BasicRestService<Company>{
 
     @Override
     public Company save(Company company) {
+        if(company.getParentCompany()!=null){
+            Company parentCompany=getEntityByID(company.getParentCompany().getId());
+            company.setParentCompany(parentCompany);
+            parentCompany.getSubCompanies().add(company);
+        }
         return companyRepository.save(company);
     }
 
@@ -27,27 +32,28 @@ public class CompanyService implements BasicRestService<Company>{
     }
 
     @Override
-    public Company getEntityByID(long id) {
-        Optional<Company> company= companyRepository.findById(id);
-        return company.orElseThrow(()-> new CompanyNotFoundException(id));
+    public Company getEntityByID(Long id) {
+        if(id==null) throw new MissingValueException("company id must not be null");
+        Optional<Company> optionlCompany= companyRepository.findById(id);
+        return optionlCompany.orElseThrow(()-> new CompanyNotFoundException(id));
     }
 
     @Override
-    public void deleteEntityByID(long id) {
-        companyRepository.deleteById(id);
+    public void deleteEntityByID(Long id) {
+        if(companyRepository.existsById(id)){
+            companyRepository.deleteById(id);
+        }else{
+            throw new CompanyNotFoundException(id);
+        }
     }
 
     @Override
     public Company updateEntity(Company company, Long id) {
         return companyRepository.findById(id).map(cmp-> {cmp.setName(company.getName());
                                                 Long parentID = null;
-                                                if (company.getParentCompany()!=null && (parentID=company.getParentCompany().getId())!=null){
-                                                    Optional<Company> parentCompany = companyRepository.findById(parentID);
-                                                    if(parentCompany.isPresent()){
-                                                        cmp.setParentCompany(parentCompany.get());
-                                                    }else{
-                                                        throw new CompanyNotFoundException(parentID);
-                                                    }
+                                                if (company.getParentCompany()!=null){
+                                                        Company parentCompany = getEntityByID(company.getParentCompany().getId());
+                                                        cmp.setParentCompany(parentCompany);
                                                 }else{
                                                     cmp.setParentCompany(null);
                                                 }
