+++ src/main/java/fi/devolon/demo/service/StationService.java
@@ -1,11 +1,13 @@
 package fi.devolon.demo.service;
 
+import fi.devolon.demo.exceptions.MissingValueException;
 import fi.devolon.demo.exceptions.StationNotFoundException;
 import fi.devolon.demo.model.Company;
 import fi.devolon.demo.model.Station;
 import fi.devolon.demo.repository.StationRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -18,6 +20,7 @@ public class StationService implements BasicRestService<Station> {
     @Override
     public Station save(Station station) {
         Company company=companyService.getEntityByID(station.getCompany().getId());
+        if (company.getStations().contains(station)) throw new DataIntegrityViolationException("Values Are Duplicate");
         company.getStations().add(station);
         station.setCompany(company);
         return stationRepository.save(station);
@@ -29,14 +32,20 @@ public class StationService implements BasicRestService<Station> {
     }
 
     @Override
-    public Station getEntityByID(long id) {
+    public Station getEntityByID(Long id) {
+        if(id==null) throw new MissingValueException("station id must not be null");
         Optional<Station> station= stationRepository.findById(id);
         return station.orElseThrow(()-> new StationNotFoundException(id));
     }
 
     @Override
-    public void deleteEntityByID(long id) {
-        stationRepository.deleteById(id);
+    public void deleteEntityByID(Long id) {
+        if(stationRepository.existsById(id)) {
+            stationRepository.deleteById(id);
+        }
+        else{
+            throw new StationNotFoundException(id);
+        }
     }
 
     @Override
@@ -49,7 +58,6 @@ public class StationService implements BasicRestService<Station> {
             company.getStations().add(stn);
             System.out.println(stn.getName());
             System.out.println(stn.getLongitude());
-
             return stationRepository.save(stn);})
                 .orElseThrow(()-> new StationNotFoundException(id));
     }
